mcq_marking/pyproject.toml
[tool.poetry]
name = "mcq-marking"
version = "0.1.0"
description = "MCQ marking and template Config system"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
pillow = ">=11.3.0,<12.0.0"
openpyxl = ">=3.1.5,<4.0.0"
opencv-python-headless = ">=4.8.0,<5.0.0"
numpy = ">=1.24.0,<2.0.0"
pika = ">=1.3.0,<2.0.0"
scikit-learn = ">=1.3.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
mcq-marking = "app.main:main"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "PIL.*",
    "openpyxl.*",
    "pika.*",
    "sklearn.*"
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"